Python/requirements.txt
flask
flask_sqlalchemy
flask_redis
pymysql
cerberus
bcrypt
itsdangerous
flask_uploads
pillow
flask_cors
openpyxl
defusedxml
flask_migrate
